Updates order status after shipping and payment
A signal was emitted to set the status to CONFIRMED after the shipping
stage, but no one listens to it. Now the status is directly modified.

Similarly, status is directly set to COMPLETED after payment. The cart
is also emptied at this stage. It was already there, but only when the
backend payment is done, and the user redirected to the "Thank you"
page. Now that logic has been copied to the confirm_payment API
function, so that the order is COMPLETED as soon as payment is made.
This is important when using tha PayPal IPN backend, since the user is
redirected back the site later than - if at all, the IPN is sent. 

Not tested!
was
already there,