BIG CHANGE
Please note that we have introduced backward incompatible changes!

First we removed SHOP_API which was maybe a good concept but not very well implemented.
We removed whole complicated backend stuff and made it waaaaaay easier and maintenable.
Every backend has an  entry in the DB and is linked to the order so you can keep track
of what was used where. Even more it adds extensibility because it was pine in the ass
to try to modify SHOP_API class or to add any backend.